Modularize glyph layer storage
