+++ crates/shift-store/src/layer/payload.rs
@@ -1,0 +1,216 @@
+use std::cell::RefCell;
+
+use super::{MAX_LAYER_PAYLOAD_BYTES, check_layer_length};
+use crate::StoreError;
+
+const ZSTD_COMPRESSION_LEVEL: i32 = 1;
+
+thread_local! {
+    static ZSTD_COMPRESSOR: RefCell<Option<zstd::bulk::Compressor<'static>>> = const { RefCell::new(None) };
+    static ZSTD_DECOMPRESSOR: RefCell<Option<zstd::bulk::Decompressor<'static>>> = const { RefCell::new(None) };
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub(crate) enum LayerPayloadCompression {
+    None,
+    ZstandardV1,
+}
+
+impl LayerPayloadCompression {
+    pub(crate) fn as_str(self) -> &'static str {
+        match self {
+            Self::None => "none",
+            Self::ZstandardV1 => "zstd.v1",
+        }
+    }
+}
+
+impl TryFrom<&str> for LayerPayloadCompression {
+    type Error = StoreError;
+
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        match value {
+            "none" => Ok(Self::None),
+            "zstd.v1" => Ok(Self::ZstandardV1),
+            value => Err(StoreError::UnsupportedLayerCompression(value.to_owned())),
+        }
+    }
+}
+
+pub(crate) struct StoredLayerPayload {
+    pub(crate) compression: LayerPayloadCompression,
+    pub(crate) bytes: Vec<u8>,
+    pub(crate) stored_byte_length: u64,
+    pub(crate) decoded_byte_length: u64,
+    pub(crate) decoded_blake3: [u8; 32],
+}
+
+pub(crate) fn compress_layer(decoded: &[u8]) -> Result<StoredLayerPayload, StoreError> {
+    check_layer_length(decoded.len() as u64)?;
+    let compressed = ZSTD_COMPRESSOR.with(|slot| {
+        let mut slot = slot.borrow_mut();
+        if slot.is_none() {
+            *slot = Some(
+                zstd::bulk::Compressor::new(ZSTD_COMPRESSION_LEVEL)
+                    .map_err(|error| StoreError::LayerCompression(error.to_string()))?,
+            );
+        }
+        slot.as_mut()
+            .expect("compressor was initialized")
+            .compress(decoded)
+            .map_err(|error| StoreError::LayerCompression(error.to_string()))
+    })?;
+    let decoded_byte_length = decoded.len() as u64;
+    let decoded_blake3 = *blake3::hash(decoded).as_bytes();
+
+    if compressed.len() < decoded.len() {
+        return Ok(StoredLayerPayload {
+            compression: LayerPayloadCompression::ZstandardV1,
+            stored_byte_length: compressed.len() as u64,
+            decoded_byte_length,
+            decoded_blake3,
+            bytes: compressed,
+        });
+    }
+
+    Ok(StoredLayerPayload {
+        compression: LayerPayloadCompression::None,
+        stored_byte_length: decoded_byte_length,
+        decoded_byte_length,
+        decoded_blake3,
+        bytes: decoded.to_vec(),
+    })
+}
+
+pub(crate) fn decompress_layer(
+    layer_id: &str,
+    stored: StoredLayerPayload,
+) -> Result<Vec<u8>, StoreError> {
+    check_layer_length(stored.stored_byte_length)?;
+    check_layer_length(stored.decoded_byte_length)?;
+    if stored.bytes.len() as u64 != stored.stored_byte_length {
+        return Err(directory_mismatch(
+            layer_id,
+            format!(
+                "stored byte length {} != actual {}",
+                stored.stored_byte_length,
+                stored.bytes.len()
+            ),
+        ));
+    }
+
+    let decoded = match stored.compression {
+        LayerPayloadCompression::None => stored.bytes,
+        LayerPayloadCompression::ZstandardV1 => {
+            let frame_length =
+                zstd::zstd_safe::find_frame_compressed_size(&stored.bytes).map_err(|code| {
+                    StoreError::LayerDecompression(zstd::zstd_safe::get_error_name(code).to_owned())
+                })?;
+            if frame_length != stored.bytes.len() {
+                return Err(directory_mismatch(
+                    layer_id,
+                    format!(
+                        "zstd frame has {frame_length} bytes, stored payload has {}",
+                        stored.bytes.len()
+                    ),
+                ));
+            }
+            let decoded_capacity = usize::try_from(stored.decoded_byte_length).map_err(|_| {
+                StoreError::LayerPayloadTooLarge {
+                    bytes: stored.decoded_byte_length,
+                    limit: MAX_LAYER_PAYLOAD_BYTES as u64,
+                }
+            })?;
+            ZSTD_DECOMPRESSOR.with(|slot| {
+                let mut slot = slot.borrow_mut();
+                if slot.is_none() {
+                    *slot = Some(
+                        zstd::bulk::Decompressor::new()
+                            .map_err(|error| StoreError::LayerDecompression(error.to_string()))?,
+                    );
+                }
+                slot.as_mut()
+                    .expect("decompressor was initialized")
+                    .decompress(&stored.bytes, decoded_capacity)
+                    .map_err(|error| StoreError::LayerDecompression(error.to_string()))
+            })?
+        }
+    };
+
+    if decoded.len() as u64 != stored.decoded_byte_length {
+        return Err(directory_mismatch(
+            layer_id,
+            format!(
+                "decoded byte length {} != actual {}",
+                stored.decoded_byte_length,
+                decoded.len()
+            ),
+        ));
+    }
+    if blake3::hash(&decoded).as_bytes() != &stored.decoded_blake3 {
+        return Err(directory_mismatch(layer_id, "decoded BLAKE3 mismatch"));
+    }
+    Ok(decoded)
+}
+
+fn directory_mismatch(layer_id: &str, detail: impl Into<String>) -> StoreError {
+    StoreError::LayerDirectoryMismatch {
+        layer_id: layer_id.to_owned(),
+        detail: detail.into(),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn incompressible_bytes_fall_back_to_none() {
+        let mut value = 0x1234_5678_u32;
+        let decoded = (0..64)
+            .map(|_| {
+                value = value.wrapping_mul(1_664_525).wrapping_add(1_013_904_223);
+                (value >> 24) as u8
+            })
+            .collect::<Vec<_>>();
+
+        let stored = compress_layer(&decoded).unwrap();
+
+        assert_eq!(stored.compression, LayerPayloadCompression::None);
+        assert_eq!(decompress_layer("layer_test", stored).unwrap(), decoded);
+    }
+
+    #[test]
+    fn compressed_bytes_reject_truncation_trailing_data_and_wrong_hash() {
+        let decoded = vec![0; 4096];
+        let mut truncated = compress_layer(&decoded).unwrap();
+        assert_eq!(truncated.compression, LayerPayloadCompression::ZstandardV1);
+        truncated.bytes.pop();
+        truncated.stored_byte_length -= 1;
+        assert!(decompress_layer("layer_test", truncated).is_err());
+
+        let mut trailing = compress_layer(&decoded).unwrap();
+        trailing.bytes.push(0);
+        trailing.stored_byte_length += 1;
+        assert!(decompress_layer("layer_test", trailing).is_err());
+
+        let mut wrong_hash = compress_layer(&decoded).unwrap();
+        wrong_hash.decoded_blake3[0] ^= 1;
+        assert!(decompress_layer("layer_test", wrong_hash).is_err());
+    }
+
+    #[test]
+    fn unsupported_compression_and_oversized_decode_are_rejected() {
+        assert!(matches!(
+            LayerPayloadCompression::try_from("zstd.v2"),
+            Err(StoreError::UnsupportedLayerCompression(_))
+        ));
+
+        let mut stored = compress_layer(&vec![0; 4096]).unwrap();
+        stored.decoded_byte_length = MAX_LAYER_PAYLOAD_BYTES as u64 + 1;
+        assert!(matches!(
+            decompress_layer("layer_test", stored),
+            Err(StoreError::LayerPayloadTooLarge { .. })
+        ));
+    }
+}
