+++ crates/shift-store/src/lib.rs
@@ -4,12 +4,11 @@ mod error;
 mod font;
 mod font_state;
 mod glyph;
-mod packed_layer;
+mod import_writer;
+mod layer;
 mod schema;
 mod source;
 mod store;
-mod stored_layer;
-mod stream_writer;
 mod types;
 mod workspace_state;
 mod write_mode;
@@ -17,13 +16,13 @@ mod write_mode;
 pub use error::StoreError;
 pub use font::FontInfo;
 pub use glyph::{GlyphRecord, NewGlyph};
-pub use packed_layer::{
+pub use import_writer::{FontImportWriter, encode_glyph_batch};
+pub use layer::{
     GLYPH_LAYER_FORMAT, GlyphLayerDirectoryEntry, MAX_LAYER_READ_BATCH_COUNT,
     MAX_LAYER_READ_BATCH_DECODED_BYTES,
 };
 pub use source::{AxisRecord, NewAxis, NewSource, SourceAxisLocation, SourceKind, SourceRecord};
 pub use store::ShiftStore;
-pub use stream_writer::{FontImportWriter, encode_glyph_batch};
 pub use types::{AxisId, GlyphId, GlyphWriteBatch, LayerBatchTiming, RevisionId, SourceId};
 pub use workspace_state::{
     Evidence, FileIdentity, SourceIdentitySnapshot, WorkspaceSourceKind, WorkspaceState,
